+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Unreleased
 
 **Added**
 
+* Added `to_configuration` and `to_configuration_primitive` to `compas_rrc.ExternalAxes` and `compas_rrc.RobotJoints`
+
 **Changed**
 
 **Fixed**